TCP metrics: Add conn.LocalAddr() as label source_ip
In the TCP metrics, add the connection.LocalAddr() as label source_ip,
analogous to the optional config setting which is also named source_ip.

This retrieves the actual Source IP used by the created TCP connection,
whether or not a source_ip was specified for a given TCP target.

This information can be helpful because hosts may have several network
interfaces which could (selected based the currently applied OS routing
tables and route metrics) be used to connect to a given target IP.